+++ trunk/ucon64/src/backup/pce-pro.h
@@ -30,8 +30,8 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 extern const st_getopt2_t pcepro_usage[];
 
 #ifdef USE_PARALLEL
-extern int pce_read_rom (const char *filename, unsigned int parport, int size);
-extern int pce_write_rom (const char *filename, unsigned int parport);
+extern int pce_read_rom (const char *filename, unsigned short parport, int size);
+extern int pce_write_rom (const char *filename, unsigned short parport);
 #endif
 
 #endif
