+++ trunk/ucon64/src/backup/sflash.h
@@ -24,10 +24,10 @@ Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 extern const st_getopt2_t sflash_usage[];
 
 #ifdef USE_PARALLEL
-extern int sf_read_rom (const char *filename, unsigned int parport, int size);
-extern int sf_write_rom (const char *filename, unsigned int parport);
-extern int sf_read_sram (const char *filename, unsigned int parport);
-extern int sf_write_sram (const char *filename, unsigned int parport);
+extern int sf_read_rom (const char *filename, unsigned short parport, int size);
+extern int sf_write_rom (const char *filename, unsigned short parport);
+extern int sf_read_sram (const char *filename, unsigned short parport);
+extern int sf_write_sram (const char *filename, unsigned short parport);
 #endif
 
 #endif
